+++ src/main/java/com/splitwizard/splitwizard/VO/OverViewVO.java
@@ -3,11 +3,13 @@ package com.splitwizard.splitwizard.VO;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Getter @Setter
 public class OverViewVO {
 
     private Integer memberId;
     private String memberName;
-    private Double memberNet;
+    private BigDecimal memberNet;
 
 }
